feat: add storage engine, bot module, infra fixes; all tests pass
- C++ Data Storage: ColumnBlock with Delta-of-Delta/Gorilla compression,
  TimeSeriesCache with 64-shard LRU eviction
- Python Feishu Bot: FeishuBot core, CommandParser, SignalPusher,
  MessageTemplates, 5 command handlers, 36 bot tests
- C++ Infra fixes: memory_pool 64-byte aligned blocks, bump allocator
  alignment, error_codes test fix, ConfigManager hot_reload,
  Metrics CAS loop, Logger JSON escaping, EventBus async worker

C++ tests: 114 passed (10 suites) | Python tests: 96 passed

+++ cpp/quant/event/events/kline_event.h
@@ -18,7 +18,7 @@ struct KlineRow {
     int32_t close_price;
     int64_t volume;
     int64_t amount;
-    int32_t vwap;
+    int64_t vwap;
 };
 static_assert(sizeof(KlineRow) == 48);
 #pragma pack(pop)
